refactor: split domain orchestrator economy pass into part module (Refs #3749 step 7)
Extract the economy pass (work/recruit/build gate helpers and the
`_runEconomyDomainPlanners` / `_appendEconomyBuildOrders` chain) out of
`domain_planner_orchestrator.dart` into a new `part` module
`domain_planner_orchestrator_economy.dart`. Behaviour-preserving move:
all symbols stay in the same library so private access and the
`runDomainPlannersWithOutcome` export are unchanged.

Shrinks the orchestrator file from 967 to 316 lines (economy part 656),
advancing the #3749 step-7 orchestrator split AC toward pass modules with
no resulting file >1000 LOC.
